lint(gui): move root fn to config module
